backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
elasticsearch==8.12.0
minio==7.2.3
openai==1.12.0
sentence-transformers==2.3.1
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
langgraph==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
celery==5.3.6
pytest==7.4.4
httpx==0.26.0
python-dotenv==1.0.0
langchain-google-genai==0.0.9
